MOB-253 Documentation on how to run a simple single-node and multi-node experiment
Added package: src/edu/umass/cs/gns/test/testclient.
It implements a stand-alone client distint from a local name server. The client and local name server
communicate using JAVA RMI interface. The class src/edu/umass/cs/gns/test/testclient/TestClient.java
is a standalone client and  includes a simple test for GNS. 

Refer to documentation and script in scripts/basic_test_local to run this test. 


git-svn-id: svn+ssh://none.cs.umass.edu/svn/mobility-first/GNS/trunk@780 0b15c3b6-4604-4ca4-ad59-2e1af984e93c

+++ scripts/basic_test_local/run_basic.sh
@@ -1,26 +1,51 @@
+gns_jar=$1
+
+if [ -z "$gns_jar" ]; then
+        echo "ERROR: No GNS jar given. Give complete path to GNS jar."
+        echo "Usage: ./run_basic.sh <Path to GNS jar>"
+        exit 2
+fi
+
+
 # run mongod
 killall -9 mongod
 rm -rf gnsdb
 mkdir -p gnsdb
 nohup mongod --dbpath=gnsdb  &
 
+
 # run name servers and local name servers
 killall -9 java
 rm -rf log
-mkdir -p log/ns0 log/ns1 log/ns2 log/lns3
+mkdir -p log/ns0 log/ns1 log/ns2 log/lns3 log/client
 
 cd log/ns0
-nohup java -ea -cp ../../../../dist/GNS.jar edu.umass.cs.gns.main.StartNameServer -id 0 -nsfile ../../node_config &
+nohup java -ea -cp $gns_jar edu.umass.cs.gns.main.StartNameServer -id 0 -nsfile ../../node_config &
 cd ../../
 
 cd log/ns1
-nohup java -ea -cp ../../../../dist/GNS.jar edu.umass.cs.gns.main.StartNameServer -id 1 -nsfile ../../node_config &
+nohup java -ea -cp $gns_jar edu.umass.cs.gns.main.StartNameServer -id 1 -nsfile ../../node_config &
 cd ../../
 
 cd log/ns2
-nohup java -ea -cp ../../../../dist/GNS.jar edu.umass.cs.gns.main.StartNameServer -id 2 -nsfile ../../node_config &
+nohup java -ea -cp $gns_jar edu.umass.cs.gns.main.StartNameServer -id 2 -nsfile ../../node_config &
 cd ../../
 
+# start rmiregistry
+killall -9 rmiregistry
+rmiregistry -J-Djava.rmi.server.useCodebaseOnly=false &
+
+# start local name server
 cd log/lns3
-nohup java -ea -cp ../../../../dist/GNS.jar edu.umass.cs.gns.main.StartLocalNameServer -id 3 -nsfile ../../node_config -experimentMode  -statFileLoggingLevel FINE -statConsoleOutputLevel FINE &
+nohup java -ea -cp $gns_jar -Djava.rmi.server.useCodebaseOnly=false -Djava.rmi.server.codebase=file:$gns_jar \
+edu.umass.cs.gns.main.StartLocalNameServer -id 3 -nsfile ../../node_config  -experimentMode \
+-statFileLoggingLevel FINE -statConsoleOutputLevel FINE &
+cd ../../
+
+# sleep so that LNS can startup and bind itself to rmiregistry
+sleep 2
+
+cd log/client
+java -ea -cp $gns_jar  -Djava.rmi.server.useCodebaseOnly=false -Djava.rmi.server.codebase=file:$gns_jar \
+ edu.umass.cs.gns.test.testclient.TestClient
 cd ../../
